+++ rtic-monotonics/CHANGELOG.md
@@ -11,6 +11,7 @@ For each category, *Added*, *Changed*, *Fixed* add new entries at the top!
 
 - Race condition in `nrf::timer`.
 - Race condition in `nrf::rtc`.
+- Add internal counter integrity check to all half-period based monotonics.
 
 ## v1.4.0 - 2023-12-04